backend/requirements.txt
-c constraints.txt
-e src/myr_plone_packages[test]

zope.testrunner

# Add required add-ons
# (Ideally add them in setup.py for myr_plone_packages)
# collective.easyform
